+++ src/parser/enrichers/class/monk/DrunkenTechnique.mjs
@@ -1,0 +1,19 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../../data/DDBEnricherData.mjs";
+
+export default class DrunkenTechnique extends DDBEnricherData {
+
+  get effects() {
+    return [
+      {
+        changes: [
+          DDBEnricherData.ChangeHelper.upgradeChange("10", 20, "system.attributes.movement.walk"),
+        ],
+        options: {
+          durationSeconds: 4,
+        },
+      },
+    ];
+  }
+
+}