+++ src/parser/enrichers/class/monk/StrideOfTheElements.mjs
@@ -5,8 +5,12 @@ export default class StrideOfTheElements extends DDBEnricherData {
 
   get effects() {
     return [
-      DDBEnricherData.ChangeHelper.upgradeChange("@attributes.movement.walk", 20, "system.attributes.movement.fly"),
-      DDBEnricherData.ChangeHelper.upgradeChange("@attributes.movement.walk", 20, "system.attributes.movement.swim"),
+      {
+        changes: [
+          DDBEnricherData.ChangeHelper.upgradeChange("@attributes.movement.walk", 20, "system.attributes.movement.fly"),
+          DDBEnricherData.ChangeHelper.upgradeChange("@attributes.movement.walk", 20, "system.attributes.movement.swim"),
+        ],
+      },
     ];
   }